Reorganized project
* Created library
* Gave `Character` its own file
* Removed all top-level methods from `main.rb` (now `skirmish`) and put them into `utilities.rb`
* Added `Utilities`
* Updated README
