+++ README.md
@@ -1,4 +1,4 @@
 # skirmish
 A single-player text-based game compatible with CircleMUD world files
 
-Run main.rb and choose a name. You will be assigned statistics and dropped into the world. Type 'commands' to see all possible commands.
+Run `ruby skirmish` and choose a name. You will be assigned statistics and dropped into the world. Type 'commands' to see all possible commands.
