+++ lib/character.rb
@@ -1,0 +1,36 @@
+class Character
+  attr_accessor :name, :height, :weight, :str, :dex, :con, :int, :wis, :cha,
+    :maxhp, :hp, :xp, :armour, :level, :state, :location
+
+  # Create the character, which by default begins at level one
+  def initialize(name, initial_level=1)
+    @name = name
+
+    size_roll = roll_dice 2, 10
+    @height = 54 + size_roll                        # inches
+    @weight = 110 + roll_dice(2, 4) * size_roll     # pounds
+
+    # roll 3d6 for attributes
+    rolls = Array.new
+    6.times {rolls.push roll_dice(3,6)}
+    rolls.sort! { |a, b| a <=> b }
+
+    @str = rolls.pop()
+    @con = rolls.pop()
+    @dex = rolls.pop()
+    @cha = rolls.pop()
+    @wis = rolls.pop()
+    @int = rolls.pop()
+    @maxhp = @con + 10
+    @hp = @maxhp
+    @xp = 0
+    @armour = @dex + 10
+    @level = initial_level
+    @location = 3001
+  end
+
+  # Character's attack each time is calculated based on a roll of 1d10 * level
+  def attack
+    roll_dice(1, 10) * @level
+  end
+end
