+++ lib/commands.rb
