+++ lib/world.rb
