+++ skirmish
@@ -1,0 +1,32 @@
+#!/usr/bin/env ruby
+
+require 'colorize'
+require_relative 'lib/world'
+require_relative 'lib/utilities'
+require_relative 'lib/commands'
+require_relative 'lib/character'
+include Utilities
+
+if __FILE__ == $0
+  fill_name_array %w(Mike Joe Alice Susan)
+
+  $world = World.new "30.wld"
+
+  until (name = prompt_user("Welcome! What is your name?")).length > 1 do
+    puts ("I'm sorry, your name must be at least two characters long.").colorize(:light_green)
+  end
+
+  $player = Character.new(name)
+
+  begin
+    Commands.stats $player
+    input = prompt_user "\nIs this acceptable (y/n)?"
+    $player = Character.new(name) if (input =~ /\An/i)
+  end until input =~ /\Ay/i
+
+  puts
+  Commands.look $player
+
+  # loop until player inputs the 'quit' command
+  loop {break if handle_input == :quit}
+end
